+++ Introverted_Developers/src/Score/Score.java
@@ -5,7 +5,7 @@ import java.util.Scanner;
 
 public class Score {
 
-    public Score_Data SD = new Score_Data();
+
 
     public void Score_A() {
 
@@ -14,7 +14,7 @@ public class Score {
         String j = sc.next();
         System.out.println("점수를 기입해주세요");
         int i = sc.nextInt();
-        SD.score(j, i);
+
     }
 
     public void Score_B() {
